backend/requirements.txt
# Flask Core
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-CORS==5.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0

# Database
psycopg2-binary==2.9.10
SQLAlchemy==2.0.36
alembic==1.18.1

# Environment
python-dotenv==1.0.1

# Security
bcrypt==5.0.0
PyJWT==2.10.1

# Integrations
mercadopago==2.3.0
requests==2.32.5

# Production Server
gunicorn==23.0.0

# Testing
pytest==9.0.2
pytest-flask==1.3.0

# Utils
blinker==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pluggy==1.6.0
Pygments==2.19.2
urllib3==2.6.3
Werkzeug==3.1.4
